assignments/requirements.txt
numpy>=1.24.0
scipy>=1.10.0
Pillow>=10.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
tqdm>=4.66.0